aider: Added `Message` and `MessageLog` classes to `message_log.py` for displaying game messages in a console display near the bottom of the game screen.
# Aider chat conversation:

USER: let's make a console display of the last game messages near the bottom of the game screen. To do this, lets create a messages_log.py with Message and MessageLog classes.

ASSISTANT: Sure, we can create a new file `message_log.py` with `Message` and `MessageLog` classes. Here's how we can structure these classes:

```python
message_log.py
<<<<<<< HEAD
=======
class Message:
    def __init__(self, text: str, fg: Tuple[int, int, int]):
        self.text = text
        self.fg = fg

class MessageLog:
    def __init__(self, x: int, width: int, height: int):
        self.messages = []
        self.x = x
        self.width = width
        self.height = height

    def add_message(self, message: Message) -> None:
        # Add a message if there is room or remove the first message
        new_msg_lines = textwrap.wrap(message.text, self.width)

        for line in new_msg_lines:
            # If the buffer is full, remove the first line to make room for the new one
            if len(self.messages) == self.height:
                del self.messages[0]

            # Add the new line as a Message object, with the text and the color
            self.messages.append(Message(line, message.fg))
>>>>>>> updated
```

This code creates a `Message` class that holds a text string and a color for the text. The `MessageLog` class holds a list of `Message` objects and has a method to add a new message to the log. If the log is full (i.e., it has reached its maximum height), the oldest message is removed to make room for the new one. The `add_message` method also wraps text so that messages longer than the width of the log are displayed correctly.