aider: Added `__iter__` method to `MessageLog` class to make it iterable.
# Aider chat conversation:

USER: 'MessageLog' object in note iterable

ASSISTANT: It seems like you're trying to iterate over a `MessageLog` object, but it's not iterable. This is likely because the `MessageLog` class does not implement an `__iter__` method, which is required for an object to be iterable in Python.

To fix this, we need to add an `__iter__` method to the `MessageLog` class. This method will allow us to iterate over the messages in the `MessageLog`.

Here is the *edit block* to add the `__iter__` method to the `MessageLog` class:

```python
message_log.py
<<<<<<< HEAD
class MessageLog:
    def __init__(self, x: int, width: int, height: int):
        self.messages = []
        self.x = x
        self.width = width
        self.height = height

    def add_message(self, message: Message) -> None:
        # Add a message if there is room or remove the first message
        new_msg_lines = textwrap.wrap(message.text, self.width)

        for line in new_msg_lines:
            # If the buffer is full, remove the first line to make room for the new one
            if len(self.messages) == self.height:
                del self.messages[0]

            # Add the new line as a Message object, with the text and the color
            self.messages.append(Message(line, message.fg))

    def get_messages(self) -> List[Message]:
        return self.messages
=======
class MessageLog:
    def __init__(self, x: int, width: int, height: int):
        self.messages = []
        self.x = x
        self.width = width
        self.height = height

    def add_message(self, message: Message) -> None:
        # Add a message if there is room or remove the first message
        new_msg_lines = textwrap.wrap(message.text, self.width)

        for line in new_msg_lines:
            # If the buffer is full, remove the first line to make room for the new one
            if len(self.messages) == self.height:
                del self.messages[0]

            # Add the new line as a Message object, with the text and the color
            self.messages.append(Message(line, message.fg))

    def get_messages(self) -> List[Message]:
        return self.messages

    def __iter__(self):
        return iter(self.messages)
>>>>>>> updated
```

This change adds the `__iter__` method to the `MessageLog` class, which returns an iterator over the messages in the `MessageLog`. This allows you to iterate over a `MessageLog` object directly, instead of having to call `get_messages` and iterate over the result.